+++ example/src/comps/flex/flex.tsx
@@ -1,13 +1,22 @@
+import { h } from 'frame3'
 import { JSX } from 'frame3/jsx-runtime'
 import './flex.css'
 
-type IFlex = JSX.HTMLAttributes<HTMLDivElement> & {
+type IFlex = JSX.HTMLAttributes<any> & {
   layout: 'h' | 'v' | 'c'
+  as?: keyof JSX.IntrinsicElements
   children?: any
 }
 
-export const Flex = ({ layout, children, className }: IFlex) => {
-  return <div class={cls('flex', className, layout)}>{children}</div>
+export const Flex = ({ as = 'div', layout, children, className, ...rest }: IFlex) => {
+  return h(
+    as,
+    {
+      class: cls('flex', className, layout),
+      ...rest,
+    },
+    [children]
+  )
 }
 
 function cls(...args: any[]) {