+++ packages/frame3/CHANGELOG.md
@@ -1,5 +1,14 @@
 # frame3
 
+## 1.1.3
+
+### Patch Changes
+
+- 修复了一些使用体验的问题
+- Updated dependencies
+  - frame3-compiler-sfc@1.1.1
+  - frame3-runtime-core@1.1.3
+
 ## 1.1.2
 
 ### Patch Changes