+++ packages/runtime-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @frame3/runtime-core
 
+## 1.1.3
+
+### Patch Changes
+
+- 修复了一些使用体验的问题
+
 ## 1.1.2
 
 ### Patch Changes